split go files
